+++ example/lib/demo/screens/color_picker/picker_sliders/wheel_diameter_slider.dart
@@ -16,7 +16,7 @@ class WheelDiameterSlider extends ConsumerWidget {
           '${ref.read(wheelDiameterPod).floor()})',
       child: ListTile(
         title: const Text('Color wheel total width'),
-        subtitle: Slider.adaptive(
+        subtitle: Slider(
           min: 150,
           max: 500,
           divisions: 500 - 150,