+++ example/lib/demo/screens/color_picker/picker_sliders/wheel_square_border_radius_slider.dart
@@ -16,7 +16,7 @@ class WheelSquareBorderRadiusSlider extends ConsumerWidget {
           '${ref.read(wheelSquareBorderRadiusPod).floor()})',
       child: ListTile(
         title: const Text('Color wheel shade square border radius'),
-        subtitle: Slider.adaptive(
+        subtitle: Slider(
           min: 0,
           max: 16,
           divisions: 16,