+++ example/lib/demo/screens/color_picker/picker_text_fields/text_field_focus_demo.dart
@@ -31,8 +31,8 @@ class _TextFieldFocusDemoState extends State<TextFieldFocusDemo> {
         const ListTile(
           title: Text('Demo of picker auto focus setting'),
           subtitle: Text('If picker auto focus is OFF or if neither keyboard '
-              'copy-paste shortcut is enabled, this field will be autofocus in '
-              'this demo. To try it, change settings, go back to main screen '
+              'copy-paste shortcut is enabled, this field will be autofocused. '
+              'To try it, change settings, go back to main screen '
               'and come back to this screen, to see the focus result.'),
         ),
         Padding(
@@ -51,7 +51,8 @@ class _TextFieldFocusDemoState extends State<TextFieldFocusDemo> {
             controller: _textController,
             decoration: InputDecoration(
               border: const OutlineInputBorder(),
-              hintText: 'If Picker autofocus is OFF field will have focus',
+              hintText: 'If Picker autofocus is OFF, '
+                  'this field will have focus',
               labelText: 'Focus demo entry',
               errorText: _errorState
                   ? "Any entry without an 'a' will trigger this error"