+++ example/lib/demo/utils/app.dart
@@ -3,22 +3,20 @@ import 'package:flutter/material.dart';
 
 /// Contains used constants for doubles, strings and colors.
 class App {
-  // This class is not meant to be instantiated or extended, this constructor
-  // prevents external instantiation and extension.
   App._();
 
   // Just a simple way to leave a trace of what version we built the Flutter
   // Web demo with inside the app. Shown on the start screen in the demo,
   // so people testing it don't have to ask. Also info for the About screen.
   static const String appName = 'FlexColorPicker';
-  static const String version = '2.6.1';
+  static const String version = '3.0.0';
   static const String packageVersion = 'FlexColorPicker package $version';
   static final Uri packageUri = Uri(
     scheme: 'https',
     host: 'pub.dev',
     path: 'packages/flex_color_picker',
   );
-  static const String flutterVersion = 'Channel stable 3.3.1';
+  static const String flutterVersion = 'Channel stable 3.3.8';
   static const String copyright = '© 2020, 2021, 2022';
   static const String author = 'Mike Rydstrom';
   static const String license = 'BSD 3-Clause License';
@@ -29,31 +27,14 @@ class App {
   // Min allowed width for the columns, if lower, column count is reduced.
   static const double minColumnWidth = 395;
   // The outline thickness of OutlinedButton and ToggleButtons
-  static const double outlineThickness = 1.5;
+  static const double outlineThickness = 1;
   // Border radius for rounded buttons
   static const double borderRadius = 50;
 
-  // Light theme colors
-  static const Color primaryLight = Color(0xFF095D9E);
-  static const Color primaryContainerLight = Color(0xFF658EAF);
-  static const Color secondaryLight = Color(0xFFDD520F);
-  static const Color secondaryContainerLight = Color(0xFFF0B03F);
-  static const Color backgroundLight = Color(0xFFFFFFFF);
-  static const Color surfaceLight = Color(0xFFFFFFFF);
+  // Theme Color
+  static const Color seedColor = Color(0xFF095D9E);
   static const Color scaffoldBackgroundLight = Color(0xFFFDFDFE);
-  static const Color onPrimaryLight = Color(0xFFFFFFFF);
-  static const Color onSecondaryLight = Color(0xFFFFFFFF);
-
-  // Dark theme colors
-  static const Color primaryDark = Color(0xFF538EBB);
-  static const Color primaryContainerDark = Color(0xFF88A7C1);
-  static const Color secondaryDark = Color(0xFFE47945);
-  static const Color secondaryContainerDark = Color(0xFFEAB04D);
-  static const Color backgroundDark = Color(0xFF161F26);
-  static const Color surfaceDark = Color(0xFF161F26);
   static const Color scaffoldBackgroundDark = Color(0xFF070A0C);
-  static const Color onPrimaryDark = Color(0xFFFFFFFF);
-  static const Color onSecondaryDark = Color(0xFFFFFFFF);
 
   // Some example custom colors for our custom picker.
   static const Color guideNewPrimary = Color(0xFF6200EE);