+++ example/lib/demo/widgets/switch_tile_tooltip.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 
 import 'maybe_tooltip.dart';
+import 'switch_list_tile_adaptive.dart';
 
 /// This simple wrapper of the SwitchListTile.adaptive widget, is used to
 /// wrap it with a tooltip and show the provided tooltip if one is provided
@@ -29,7 +30,7 @@ class SwitchTileTooltip extends StatelessWidget {
     return MaybeTooltip(
       condition: tooltipEnabled,
       tooltip: tooltip,
-      child: SwitchListTile.adaptive(
+      child: SwitchListTileAdaptive(
         value: value,
         onChanged: onChanged,
         title: title,